refactor: modularize dice pool editor
